requirements.txt
#dlordinal==2.4.1
dlordinal @ git+https://github.com/stefanahaas41/dlordinal@enhancement/129-unimodal-copoc
#MAPIE==1.1.0
MAPIE @ git+https://github.com/lucatk/MAPIE@rps-score
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
skorch==1.1.0
streamlit==1.44.1
tornado==6.4.2
watchdog==6.0.0
matplotlib==3.10.1
matplot2tikz==0.4.1
medmnist==3.0.2
pillow~=11.2.1
highlight_text==0.2
dotenv==0.9.9
tqdm==4.67.1
seaborn==0.13.2

--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.0+cu128; sys_platform == "linux"
torchvision==0.22.0+cu128; sys_platform == "linux"

torch==2.7.0; sys_platform == "darwin"
torchvision==0.22.0; sys_platform == "darwin"
